+++ ledger/applications.go
@@ -34,7 +34,7 @@ type logicLedger struct {
 
 type cowForLogicLedger interface {
 	Get(addr basics.Address, withPendingRewards bool) (basics.AccountData, error)
-	GetCreatableID(groupIdx int) basics.CreatableIndex
+	GetCreatableID(groupIdx byte) basics.CreatableIndex
 	GetCreator(cidx basics.CreatableIndex, ctype basics.CreatableType) (basics.Address, bool, error)
 	GetKey(addr basics.Address, aidx basics.AppIndex, global bool, key string, accountIdx uint64) (basics.TealValue, bool, error)
 	BuildEvalDelta(aidx basics.AppIndex, txn *transactions.Transaction) (transactions.EvalDelta, error)
@@ -98,7 +98,7 @@ func (al *logicLedger) Authorizer(addr basics.Address) (basics.Address, error) {
 	return addr, nil
 }
 
-func (al *logicLedger) GetCreatableID(groupIdx int) basics.CreatableIndex {
+func (al *logicLedger) GetCreatableID(groupIdx byte) basics.CreatableIndex {
 	return al.cow.GetCreatableID(groupIdx)
 }
